requirements.txt
opencv-python>=4.5.0
Flask>=3.0.0
Pillow>=9.0.0
pandas>=1.3.0
numpy>=1.20.0
requests>=2.28.0
beautifulsoup4>=4.11.0
Werkzeug>=2.0.0
matplotlib>=3.3.0
